Add files via upload
This assignment is designed to demonstrate engagement with a broad set of topics related
to object-oriented programming. Among these: knowledge of both implementing and using queues, ability to
abstract a class definition by having objects such as lists and queues inside other objects, testing via
Java Junit tests, and to find and use online documentation about the built-in Java ArrayList and
LinkedList classes. In addition, this project demonstrates the practice of dividing bigger
tasks into smaller components, and then implementing those smaller components, which is an
important skill in computer programming.